% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_helpers.R
\name{center_corr}
\alias{center_corr}
\title{Correct the center of a chron}
\usage{
center_corr(center, width)
}
\arguments{
\item{center}{The center of the chron}

\item{width}{The width of the chron}
}
\value{
A vector
}
\description{
This function corrects the calculated center of a chron
}
\keyword{internal}
