% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_convert_to_chron.R
\name{convert_to_chron}
\alias{convert_to_chron}
\title{Prepare an existing data set for plotting}
\usage{
convert_to_chron(data, region, name, start, end, level, add, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{region}{A character string with the column name of the regions/sections
in the plot.}

\item{name}{A character string with the column name of the names of the
chronological units. Must be a character string.}

\item{start}{A character string with the column name of the start dates of
the chronological units.}

\item{end}{A character string with the column name of the end date of the
chronological units.}

\item{level}{A character string with the column name of the level of the
chronological unit.}

\item{add}{A character string with the column name of the columns which
signals whether the chronological units within a geographical area should
be drawn separately.}

\item{...}{Additional columns to include in the data set and/or additional
arguments passed to \code{\link[dplyr]{rename()}}.}
}
\value{
A tibble with chronological data ready-to-use for plotting with
  \code{\link[chronochrt]{plot_chronochrt}}.
}
\description{
Convert an existing data set in a ready-to-plot data set.
}
\details{
Additional columns useful for plotting might e.g. specify the x and y
position of the names to place them at an arbitrary spot
}
