% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_geom.R
\name{corr_xmax}
\alias{corr_xmax}
\title{Determine maximum x value of parallel chrons}
\usage{
corr_xmax(start, end, xmax)
}
\arguments{
\item{start, end}{Start and end of the chrons.}

\item{xmax}{The upper x value of the chrons.}
}
\value{
A vector
}
\description{
This function determines the maximum x value of parallel chrons
}
\keyword{internal}
