% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_import.R
\name{import_chron_csv}
\alias{import_chron_csv}
\title{Import chronological information from .csv files}
\usage{
import_chron_csv(path, region, name, start, end, level, add, delim, ...)
}
\arguments{
\item{path}{the path of the file to be imported.}

\item{region}{specifies the colum of the geografical area/Level 1
(overarching) headline.}

\item{name}{specifies the colum of the name of the time period.}

\item{start}{specifies the colum of the start date (can be negative).}

\item{end}{specifies the colum of the end date (can be negative).}

\item{level}{specifies the colum of sublevel.}

\item{add}{= TRUE or FALSE. A binary operator indicating whether data within
"region" should be displayed as separate column. Defaults to FALSE.}

\item{delim}{Single character used to separate fields within a record}

\item{...}{Additional arguments inherited from \code{read_csv()}.}
}
\value{
A tibble containing the desired chronological information.
}
\description{
This function imports and converts chronological information saved in
\code{.csv} into a ready-to-plot data set. Missing values will be substituted
by "!".
}
\examples{



}
\author{
Thomas Rose
}
