% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochRt_helpers.R
\name{subchron_count}
\alias{subchron_count}
\title{Calcuate number of subchrons}
\usage{
subchron_count(left, right)
}
\arguments{
\item{left}{Left argument in evaluation}

\item{right}{Right argument in evaluation}
}
\value{
A vector
}
\description{
This function is used to calculate for each chron the number of parallel chrons.
}
\keyword{internal}
