% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_theme.R
\name{theme_chronochrt}
\alias{theme_chronochrt}
\title{The ChronochRt ggplot theme}
\usage{
theme_chronochrt(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
A ggplot-theme
}
\description{
This is a theme to provide a ready-to-use layout of chronological charts. See
\code{\link[ggplot2]{theme}} for how to modify selected elements of it.
}
