% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochRt_helpers.R
\name{width_corr}
\alias{width_corr}
\title{Correct the width of a chron}
\usage{
width_corr(center, width)
}
\arguments{
\item{center}{The center of the chron}

\item{width}{The width of the chron}
}
\value{
A vector
}
\description{
This function corrects the calculated width of a chron
}
\keyword{internal}
