% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_add_chron.R
\name{add_chron}
\alias{add_chron}
\title{Create or add Chronological data}
\usage{
add_chron(
  data,
  region,
  name,
  start,
  end,
  level = 1,
  add = FALSE,
  new_table = FALSE,
  ...
)
}
\arguments{
\item{data}{A data set with chronological data. Must not be provided if
\code{new_table = FALSE}.}

\item{region}{A character string or character vector with the title(s) of the
section(s).}

\item{name}{A character string or character vector with the name(s) of the
chronological unit(s).}

\item{start}{A number or a vector with the start date(s) of the chronological
unit(s). Use negative values for BCE dates. See Details how to handle
unsecure start dates.}

\item{end}{A number or a vector with the end date(s) of the chronological
unit(s). Use negative values for BCE dates. See Details how to handle
unsecure end dates.}

\item{level}{A whole number or numeric vector of whole numbers (i.e. 1, 2, 3,
...) with the level(s) of the chronological unit(s). The default is
\code{1}, i.e. the top unit.}

\item{add}{A logical value (\code{TRUE} or \code{FALSE}) or a logical vector
signaling whether the chronological units within a georaphical area should
be drawn separately (\code{TRUE}) or not (\code{FLASE}, the default).}

\item{new_table}{Logical operator. If \code{TRUE}, a new data set will be
created. If \code{FALSE}, the default, the input will be added to an
existing data set.}

\item{...}{Further arguments or columns to include in or additional arguments
passed to \code{\link[tibble]{tibble()}} or
\code{\link[tibble]{add_row()}}.}
}
\value{
A tibble with chronological data ready-to-use for plotting with
  \code{\link[chronochrt]{plot_chronochrt}}.
}
\description{
This function either creates a data set with chronological data which can be
directly used for plotting or it adds chronological data to such a data set.
}
\details{
If the input is in the same order like the arguments, the arguments does not
need to be explicitly named. Values can be provided as one number or one
character string, if they are the same for all other data. If not, they must
be provided as vectors with equal lengths.

\code{start} and \code{end} of neighbouring chronological units must be fit
to get a good plotting results, as well as the start date of e.g. the top
chronological unit and its respective oldest sub-units. Dates in BCE must
provided as negative data. Currently, only years can be handled (i.e. 2020
but not 20.20.2020).

Is \code{start} and \code{end} dates are not certain or the change between
chronological units is regarded a period, dates must be given as character
string in the format \code{"1000/2000"}. The order does not matter. But also
here, be consistent in order and in matching \code{start} and \code{end}
dates to avoid unclean borders in the plot.

The \code{level} indicates the position of the chronological unit.
\code{level = 1} denotes a top chronological unit (e.g. Ha), a sub-unit (e.g.
Ha B) is \code{level = 2}, a sub-sub-unit (e.g. Ha B1) \code{level = 3} etc.
The parameter \code{add} indicates whether the respective chronological
unit(s) should be plotted in the same or an additional column. This might be
useful if competing chronologies in one region exist (e.g. short and long
chronologies). See the vignette for a detailed explanation how the parameters
\code{level} and \code{add} work.

Additional columns might be useful to e.g.  specify the x and y position of
the names of the chronological units to place them at an arbitrary spot.
}
\examples{


}
