% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_arrange_regions.R
\name{arrange_regions}
\alias{arrange_regions}
\title{Arranging the regions (sections) of a chronological chart}
\usage{
arrange_regions(data, order)
}
\arguments{
\item{data}{A data set with chronological data and a cloumn named "region".}

\item{order}{A character vector with desired order of the region/section
titles. Each title must be given only once.}

\item{region}{A character string with the column name of the regions/sections}
}
\value{
A tibble with chronological data ready-to-use for plotting with
  \code{\link[chronochrt]{plot_chronochrt}}.
}
\description{
This function ensures that the regions/sections of a chronological chart are
arranged different from an alphabetal order.
}
