% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_import.R
\name{import_chron_csv}
\alias{import_chron_csv}
\title{Import chronological information from .csv files}
\usage{
import_chron_csv(path, delim, ...)
}
\arguments{
\item{path}{the path of the file to be imported.}

\item{delim}{A character string with the separator}

\item{...}{Additional arguments inherited from \code{\link[readr]{read_csv}}.}
}
\value{
A tibble containing the desired chronological information.
}
\description{
This function imports and converts chronological information saved in
\code{.csv} into a ready-to-plot data set. Missing values will be substituted
by "!".
}
\keyword{internal}
