% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronochrt_import.R
\name{import_chron_delim}
\alias{import_chron_delim}
\title{Import chronological information from other delimited files}
\usage{
import_chron_delim(path, delim, ...)
}
\arguments{
\item{path}{the path of the file to be imported.}

\item{delim}{A character string with the separator}

\item{...}{Additional arguments inherited from \code{\link[readr]{read_delim}} and
\code{\link[readr]{read_table}}.}
}
\value{
A tibble containing the desired chronological information.
}
\description{
This function imports and converts chronological information saved as tabular
data into a ready-to-plot data set. Missing values will be substituted by
"!".
}
\keyword{internal}
